BE: API Elimina test case caricato e test suite caricato

+++ be/sipp/src/main/java/it/reply/genesis/model/repository/FileSystemRepository.java
@@ -12,7 +12,7 @@ public interface FileSystemRepository extends JpaRepository<FileSystemVO, Long> 
 
   public List<FileSystemVO> findByScopeAndIdRef(FileSystemScope scope, Long idRef);
 
-  public int deleteAllByScopeAndIdRef(FileSystemScope scope, Long idRef);
+  public long deleteAllByScopeAndIdRef(FileSystemScope scope, Long idRef);
 
   public Optional<FileSystemVO> findByScopeAndIdRefAndPath(FileSystemScope scope, long idRef, String path);
